+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-# 1.0.x
+# 1.0.3
 
+- Arguments without spaces can now be specified without double-quotes
+- Added tests for command parser input patterns
 - Added `CHANGELOG.md` and `README.md` for documentation